feat: initialize repository via Create-React-App
Create-React-App is used here because it's a common build setup
for React users.

Some of the starter files have been removed, as the focus should
be on showcasing our components and less on the boilerplate of
CRA.

The package manager chosen is yarn, as it's what we use for the
react-bootstrap repository.

The current revision of this example uses beta release 9, so it
felt fitting to showcase the newest component support (Toasts).